backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic
pydantic-settings
python-multipart
httpx
python-jose[cryptography]
passlib[bcrypt]
email-validator
google-generativeai
